Make WorldKit work with newer Kotlin, non-experimental coroutines, newer LWJGL and newer JVM.

+++ src/main/kotlin/com/grimfox/gec/ui/nvgproxy/nvgproxy.kt
@@ -107,7 +107,7 @@ fun nvgSave(ctx: Long) {
 }
 
 fun nvgBeginFrame(ctx: Long, windowWidth: Int, windowHeight: Int, devicePixelRatio: Float) {
-    NanoVG.nvgBeginFrame(ctx, windowWidth, windowHeight, devicePixelRatio)
+    NanoVG.nvgBeginFrame(ctx, windowWidth.toFloat(), windowHeight.toFloat(), devicePixelRatio)
 }
 
 fun nvgEndFrame(ctx: Long) {
